Create files for platform specific code for HelpButton
