+++ src/components/SidePanel/HelpComponents/HelpButton/HelpButtonBase.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import Icon from '@components/Icon';
+import * as Expensicons from '@components/Icon/Expensicons';
+import {PressableWithoutFeedback} from '@components/Pressable';
+import Tooltip from '@components/Tooltip';
+import useLocalize from '@hooks/useLocalize';
+import useSidePanel from '@hooks/useSidePanel';
+import useTheme from '@hooks/useTheme';
+import useThemeStyles from '@hooks/useThemeStyles';
+import type HelpButtonProps from './types';
+
+function HelpButton({style}: HelpButtonProps) {
+    const styles = useThemeStyles();
+    const theme = useTheme();
+    const {translate} = useLocalize();
+    const {openSidePanel, shouldHideHelpButton} = useSidePanel();
+
+    if (shouldHideHelpButton) {
+        return null;
+    }
+
+    return (
+        <Tooltip text={translate('common.help')}>
+            <PressableWithoutFeedback
+                accessibilityLabel={translate('common.help')}
+                style={[styles.flexRow, styles.touchableButtonImage, style]}
+                onPress={openSidePanel}
+            >
+                <Icon
+                    src={Expensicons.QuestionMark}
+                    fill={theme.icon}
+                />
+            </PressableWithoutFeedback>
+        </Tooltip>
+    );
+}
+
+HelpButton.displayName = 'HelpButtonBase';
+
+export default HelpButton;
