+++ src/components/SidePanel/HelpComponents/HelpButton/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import HelpButtonBase from './HelpButtonBase';
+import type HelpButtonProps from './types';
+
+function HelpButton({style}: HelpButtonProps) {
+    return <HelpButtonBase style={style} />;
+}
+
+HelpButton.displayName = 'HelpButton';
+
+export default HelpButton;
