+++ src/components/SidePanel/HelpComponents/HelpButton/types.ts
@@ -1,0 +1,7 @@
+import type {StyleProp, ViewStyle} from 'react-native';
+
+type HelpButtonProps = {
+    style?: StyleProp<ViewStyle>;
+};
+
+export default HelpButtonProps;
